fix: data blocks causing issues with hyphenated strings (#31)
Closes #30 

The `data` terraform token was not being treated as a top level object
because we didn't have specific rules we wanted to enforce in those
blocks.

This led to a weird scenario that caused all subsequent strings
following a data block to be check against TF001, no hyphens in resource
names, which led to false positive errors.

Blocks like

```		
data "google_project" "gad" {
  project_id = module.imports.data.action_dispatcher.components.webhook.stage.project_id
}

module "common" {
  domains = ["action-dispatcher-stage.ghss.joonix.net"]
}
```

Would error on `domains = ["action-dispatcher-stage.ghss.joonix.net"]`
which is valid terraform and should not have been flagged.